refactor(types): port upload types to explicit imports
