+++ src/api/uploads/BaseMultiput.js
@@ -4,6 +4,7 @@
  * @author Box
  */
 import BaseUpload from './BaseUpload';
+import type { MultiputConfig } from '../../common/types/upload';
 
 const DEFAULT_MULTIPUT_CONFIG: MultiputConfig = {
     digestReadahead: 5, // How many parts past those currently uploading to precompute digest for
