+++ src/elements/content-uploader/nameCellRenderer.js
@@ -5,6 +5,7 @@
 
 import React from 'react';
 import IconName from './IconName';
+import type { UploadItem } from '../../common/types/upload';
 
 type Props = {
     rowData: UploadItem,
